requirements.txt
streamlit
PyMuPDF
openai
langchain
langchain_openai
langchain_experimental 
langchainhub
pinecone-client
tiktoken
python-dotenv
docx2txt
requests
numpy
pandas
pypdf